Use local file loading method in FileLoader
Before, to switch formats you had to extend and override a bunch of methods:

* `\Illuminate\Config\FileLoader::load`
* `\Illuminate\Config\FileLoader::mergeEnvironment`
* `\Illuminate\Config\FileLoader::getRequire`

This is a pain for obvious reasons, I may want to use json format, but I might want the Load method to work as it does in the framework. with this change, the only method that needs overriding is `\Illuminate\Config\FileLoader::getRequire`.