Database: reduce overhead caching mutators.
Models have quite a lot of methods even by default, and may have more
themselves.  With a lot of models, a regular expression on each adds up
fairly quickly.

This cuts it down to only those likely to match.

It would be more ideal to use preg_grep if it had a flag to produce only
the captures in the result array.