Unify database testing traits with a new FreshDatabase trait.
This trait is based on work by @adamwathan on a better testing workflow
when interacting with databases. In general, the goal is to have a
single trait that “just works” out of the box regardless of whether you
are using an in-memory SQLite database or a conventional MySQL /
Postgres database when testing. The database will be migrated to make
sure it is up to date and then a transaction will be started to make
any records added during the test temporary. Like the database
transaction trait, the current transaction will be rolled back at the
end of each test.