Remove unused HttpException
This exception was added in a previous PR to provide a more clear exception when crawling errors were presented. The crawling functionality has thus been removed from Laravel but it seems this exception was forgotten. It currently also conflicts with the inherited class which was made final in the latest PHPUnit 8 release.

Previous PR: https://github.com/laravel/framework/pull/10230