Added warning when not renaming the Redis facade
When using `phpredis` as the driver, when the Redis Facade is aliased, the connector will create a new Facade instance instead of the expected Redis Client instance. If the Facade is detected, an exception will be thrown telling the developer to rename the Facade to avoid collisions with the expected class.