Fixed increment bug that changes expiration to forever (#32875)
A bug was introduced recently when changing:

```
if ($existing = $this->get($key)) {
```

to

```
if (! isset($this->storage[$key])) {
```

When attempting to increment a value that starts from zero (0), before it worked correctly in that it was simply returning true in that a value was set. However, now when `$this->get($key)` returns 0 since we are incrementing from 0, it evaluates as false and thus changes expiration to "forever". This causes issues for testing rate limiting because nothing ever "expires".