Minor isClean docblock fix (#37807)
Hi,

It seems there's a minor mistake on the docblock that can be confusing because most IDE will show it on suggestions. Per documentation and code review, `isClean` will check if the model is clean when no arguments are passed, **or** if any of the passed attributes have been changed.