Sorting possible models and events (#48249)
With Laravel Prompts and the new dropdown combobox, you have the opportunity to open the dropdown and see a list of all possible models and events for the various commands. Currently the models and events are not sorted in any logical order that is readily visible. 

This PR sorts the models and events by name to ensure the dropdown looks nicer.