[11.x] Consider after_commit config in SyncQueue (#51071)
* Consider after_commit config in SyncQueue

AfterCommit support was added to the SyncQueue in #48860. However, it only works if the job in question implements the ShouldQueueAfterCommit interface or sets the afterCommit property - it does not work if after_commit is set to true in the sync driver's configuration. This makes the behavior inconsistent, which is what this commit solves by also considering the after_commit config.

* Update SyncQueue.php

* Update SyncQueue.php

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>